refactor(executor): move prompt-building functions to separate file
- Move BuildPrompt(), buildRetryPrompt(), buildSelfReviewPrompt(), appendResearchContext() to prompt_builder.go
- Zero coupling to orchestration loop - functions only build and return strings
- All tests passing, build verified

(GH-1137)
