+++ .agent/tasks/gh-1137.md
@@ -1,0 +1,10 @@
+# GH-1137
+
+**Created:** 2026-02-14
+
+## Problem
+
+Move `BuildPrompt()`, `buildRetryPrompt()`, `buildSelfReviewPrompt()`, `appendResearchContext()` — the 4 prompt-building functions. Zero coupling to the orchestration loop; they build strings and return them. Verify with `go build`.
+
+## Acceptance Criteria
+
