+++ .agent/tasks/q-1771018269.md
@@ -1,0 +1,15 @@
+# Q-1771018269
+
+**Created:** 2026-02-13
+
+## Problem
+
+Answer this question about the codebase. DO NOT make any changes, only read and analyze.
+
+Question: How is it going?
+
+IMPORTANT: Be concise. Limit your exploration to 5-10 files max. Provide a brief, direct answer.
+If the question is too broad, ask for clarification instead of exploring everything.
+
+## Acceptance Criteria
+
